requirements.txt
streamlit
opencv-python
numpy
pandas
python-dotenv
bcrypt
email-validator
reportlab
Pillow
twilio
